trying to get solid_cache_specs run again
