+++ card/spec/spec_helper.rb
@@ -40,6 +40,45 @@ Spork.each_run do
   # This code will be run each time you run your specs.
 end
 
+class Card
+
+  def format_with_abstract_set set, format_type=:html
+    binding.pry
+    if Object.send(:const_defined?, "Card::Set::All::TestSet")
+      Object.send :remove_const, Card::Set::Self::TestSet
+    end
+    eval %(
+      class ::Card
+        module Set
+          class Self
+            module TestSet
+              extend Card::Set
+              include_set #{set}
+            end
+          end
+        end
+      end
+    )
+    binding.pry
+    singleton_class.send :include, Card::Set::Self::TestSet
+    format = format format_type
+    format_class = Card::Format.format_class_name format_type
+    binding.pry
+    format.singleton_class.send :include, set.const_get(format_class)
+    yield(format)
+  end
+
+  def format_with_set set, format_type=:html
+    binding.pry
+
+    singleton_class.send :include, set
+    format = format format_type
+    format_class = Card::Format.format_class_name format_type
+    format.singleton_class.send :include, set.const_get(format_class)
+    yield(format)
+  end
+end
+
 SpecLoader.card_helper
 
 class ActiveSupport::Logger
